+++ src/esm/Bounds.d.ts
@@ -15,7 +15,8 @@
  * @modified 2025-04-26 Added static method `Bounds.computeFromBoundsSet` to calculate containing bounds for a set of bounding boxes.
  * @modified 2025-10-17 Added the methods `Bounds.getSouthWestPoint`, `getNorthWestPoint`, `getNorthEastPoint` and `getSouthEastPoint`.
  * @modified 2025-10-18 Added method `Bounds.containsVert(XYCoords)`.
- * @version  1.10.0
+ * @modified 2025-10-28 Added the `Bounds.getWidth()` and `Bounds.getHeight()`.
+ * @version  1.11.0
  **/
 import { Polygon } from "./Polygon";
 import { XYCoords, IBounds, XYDimension } from "./interfaces";
@@ -75,6 +76,8 @@ export declare class Bounds implements IBounds, XYDimension {
      * @memberof Bounds
      **/
     containsVert(vert: XYCoords): boolean;
+    getWidth(): number;
+    getHeight(): number;
     /**
      * Get the center point of the north bound.
      *